finished restructuring & added readme content
