+++ Options/options.html
@@ -8,7 +8,7 @@
 	</head>
 	<body>
 		<label>
-			<input type="checkbox" id="site_identifier" checked/>
+			<input type="checkbox" id="site_identifier"/>
 			Site Identifier
 		</label>
         <span id="status"></span>
