+++ index.css
@@ -35,6 +35,19 @@ button:hover{
 	cursor: pointer;
 }
 
+button img{
+	width: 25px;
+	margin-right: 5px;
+}
+
+button.jira_injection_btn{
+	padding: 0;
+	display: flex;
+	align-items: center;
+	padding-right: 5px;
+	border-radius: 5px;
+}
+
 .flex {
 	display: flex;
 	align-items: center;
